[CoronaStats] Bridge and world thing merged
