+++ bundles/org.openhab.binding.coronastats/src/main/java/org/openhab/binding/coronastats/internal/dto/CoronaStats.java
@@ -12,22 +12,12 @@
  */
 package org.openhab.binding.coronastats.internal.dto;
 
-import static org.openhab.binding.coronastats.internal.CoronaStatsBindingConstants.*;
-
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.Collections;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Set;
 
 import com.google.gson.annotations.SerializedName;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.smarthome.core.library.types.DateTimeType;
-import org.eclipse.smarthome.core.types.State;
 
 /**
  * The {@link CoronaStats} class is internal CoronaStats structure.
@@ -36,22 +26,11 @@ import org.eclipse.smarthome.core.types.State;
  */
 @NonNullByDefault
 public class CoronaStats {
-    private final Date created = new Date();
-
     @SerializedName("data")
     private @Nullable Set<CoronaStatsCountry> countries;
 
     @SerializedName("worldStats")
     private @Nullable CoronaStatsWorld world;
-
-    public Map<String, State> getChannelsStateMap() {
-        Map<String, State> map = new HashMap<>();
-
-        ZonedDateTime zoned = ZonedDateTime.ofInstant(created.toInstant(), ZoneId.systemDefault());
-        map.put(CHANNEL_REFRESHED, new DateTimeType(zoned));
-
-        return Collections.unmodifiableMap(map);
-    }
 
     public @Nullable CoronaStatsCountry getCountry(String countryCodeKey) {
         final Set<CoronaStatsCountry> localCountries = countries;
