+++ bundles/org.openhab.binding.coronastats/src/main/java/org/openhab/binding/coronastats/internal/handler/CoronaStatsWorldHandler.java
@@ -12,55 +12,208 @@
  */
 package org.openhab.binding.coronastats.internal.handler;
 
+import java.net.URI;
+import java.util.Set;
+import java.net.SocketTimeoutException;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonSyntaxException;
+
 import org.eclipse.jdt.annotation.NonNullByDefault;
+import org.eclipse.jdt.annotation.Nullable;
+import org.eclipse.jetty.client.HttpClient;
+import org.eclipse.jetty.client.HttpResponse;
+import org.eclipse.jetty.client.api.Request;
+import org.eclipse.jetty.client.api.Result;
+import org.eclipse.jetty.client.util.BufferingResponseListener;
+import org.eclipse.jetty.http.HttpMethod;
+import org.eclipse.smarthome.core.thing.Bridge;
+import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
 import org.eclipse.smarthome.core.thing.ThingStatusDetail;
+import org.eclipse.smarthome.core.thing.binding.BaseBridgeHandler;
+import org.eclipse.smarthome.core.thing.binding.ThingHandler;
+import org.eclipse.smarthome.core.types.Command;
+import org.eclipse.smarthome.core.types.RefreshType;
+import org.openhab.binding.coronastats.internal.CoronaStatsPollingException;
+import org.openhab.binding.coronastats.internal.config.CoronaStatsWorldConfiguration;
 import org.openhab.binding.coronastats.internal.dto.CoronaStats;
 import org.openhab.binding.coronastats.internal.dto.CoronaStatsWorld;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * The {@link CoronaStatsCountryHandler} is the handler for country thing
+ * The {@link CoronaStatsWorldHandler} is the handler for bridge thing
  *
  * @author Johannes Ott - Initial contribution
  */
 @NonNullByDefault
-public class CoronaStatsWorldHandler extends CoronaStatsThingHandler {
+public class CoronaStatsWorldHandler extends BaseBridgeHandler {
+    private static final String CORONASTATS_URL = "https://corona-stats.online/?format=json";
 
     private final Logger logger = LoggerFactory.getLogger(CoronaStatsWorldHandler.class);
 
-    public CoronaStatsWorldHandler(Thing thing) {
-        super(thing);
+    private CoronaStatsWorldConfiguration worldConfig = new CoronaStatsWorldConfiguration();
+    private @Nullable ScheduledFuture<?> pollingJob;
+    private @Nullable CoronaStats coronaStats;
+    private final Set<CoronaStatsCountryHandler> countryListeners = ConcurrentHashMap.newKeySet();
+    private final HttpClient client;
+    private final Gson gson = new Gson();
+
+    public CoronaStatsWorldHandler(Bridge bridge, HttpClient client) {
+        super(bridge);
+        this.client = client;
+    }
+
+    @Override
+    public void handleCommand(ChannelUID channelUID, Command command) {
+        if (command instanceof RefreshType) {
+            final CoronaStats localCoronaStats = coronaStats;
+            if (localCoronaStats != null) {
+                notifyOnUpdate(localCoronaStats);
+            }
+        }
     }
 
     @Override
     public void initialize() {
-        logger.debug("Initializing Corona Stats world handler");
+        logger.debug("Initializing Corona Stats bridge handler");
+        worldConfig = getConfigAs(CoronaStatsWorldConfiguration.class);
 
-        CoronaStatsBridgeHandler handler = getBridgeHandler();
-        if (handler == null) {
-            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR, "Bridge handler missing");
+        if (worldConfig.isValid()) {
+            startPolling();
+            updateStatus(ThingStatus.UNKNOWN);
         } else {
-            updateStatus(ThingStatus.ONLINE);
+            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                    "Refresh interval has to be at least 15 minutes.");
         }
     }
 
     @Override
     public void dispose() {
-        logger.debug("CoronaStats world handler disposes.");
+        logger.debug("Handler disposed.");
+        stopPolling();
     }
 
-    public void notifyOnUpdate(CoronaStats coronaStats) {
-        CoronaStatsWorld world = coronaStats.getWorld();
-        if (world == null) {
-            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR, "World stats not found");
-            return;
+    private void startPolling() {
+        final ScheduledFuture<?> localPollingJob = this.pollingJob;
+        if (localPollingJob == null || localPollingJob.isCancelled()) {
+            logger.debug("Start polling.");
+            pollingJob = scheduler.scheduleWithFixedDelay(this::poll, 0, worldConfig.refresh, TimeUnit.MINUTES);
         }
+    }
 
-        updateStatus(ThingStatus.ONLINE);
+    private void stopPolling() {
+        final ScheduledFuture<?> localPollingJob = this.pollingJob;
+        if (localPollingJob != null && !localPollingJob.isCancelled()) {
+            logger.debug("Stop polling.");
+            localPollingJob.cancel(true);
+            pollingJob = null;
+        }
+    }
 
-        world.getChannelsStateMap().forEach(this::updateState);
+    private void poll() {
+        logger.debug("Polling");
+        requestRefresh().handle((resultCoronaStats, pollException) -> {
+            if (resultCoronaStats == null) {
+                if (pollException == null) {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR);
+                } else {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR,
+                            pollException.getMessage());
+                }
+            } else {
+                updateStatus(ThingStatus.ONLINE);
+                notifyOnUpdate(resultCoronaStats);
+            }
+
+            return null;
+        });
+    }
+
+    private CompletableFuture<@Nullable CoronaStats> requestRefresh() {
+        CompletableFuture<@Nullable CoronaStats> f = new CompletableFuture<>();
+        Request request = client.newRequest(URI.create(CORONASTATS_URL));
+
+        request.method(HttpMethod.GET).timeout(2000, TimeUnit.SECONDS).send(new BufferingResponseListener() {
+            @NonNullByDefault({})
+            @Override
+            public void onComplete(Result result) {
+                final HttpResponse response = (HttpResponse) result.getResponse();
+                if (result.getFailure() != null) {
+                    Throwable e = result.getFailure();
+                    if (e instanceof SocketTimeoutException || e instanceof TimeoutException) {
+                        f.completeExceptionally(new CoronaStatsPollingException("Request timeout", e));
+                    } else {
+                        f.completeExceptionally(new CoronaStatsPollingException("Request failed", e));
+                    }
+                } else if (response.getStatus() != 200) {
+                    f.completeExceptionally(new CoronaStatsPollingException(getContentAsString()));
+                } else {
+                    try {
+                        CoronaStats coronaStatsJSON = gson.fromJson(getContentAsString(), CoronaStats.class);
+                        f.complete(coronaStatsJSON);
+                    } catch (JsonSyntaxException parseException) {
+                        logger.error("Parsing failed: {}", parseException.getMessage());
+                        f.completeExceptionally(new CoronaStatsPollingException("Parsing of response failed"));
+                    }
+                }
+            }
+        });
+
+        return f;
+    }
+
+    @Override
+    public void childHandlerInitialized(ThingHandler childHandler, Thing childThing) {
+        if (childHandler instanceof CoronaStatsCountryHandler) {
+            logger.debug("Register thing listener.");
+            final CoronaStatsCountryHandler listener = (CoronaStatsCountryHandler) childHandler;
+            if (countryListeners.add(listener)) {
+                final CoronaStats localCoronaStats = coronaStats;
+                if (localCoronaStats != null) {
+                    listener.notifyOnUpdate(localCoronaStats);
+                }
+            } else {
+                logger.warn("Tried to add listener {} but it was already present. This is probably an error.",
+                        childHandler);
+            }
+        }
+    }
+
+    @Override
+    public void childHandlerDisposed(ThingHandler childHandler, Thing childThing) {
+        if (childHandler instanceof CoronaStatsCountryHandler) {
+            logger.debug("Unregister thing listener.");
+            if (!countryListeners.remove((CoronaStatsCountryHandler) childHandler)) {
+                logger.warn("Tried to remove listener {} but it was not registered. This is probably an error.",
+                        childHandler);
+            }
+        }
+    }
+
+    public void notifyOnUpdate(@Nullable CoronaStats newCoronaStats) {
+        if (newCoronaStats != null) {
+            coronaStats = newCoronaStats;
+
+            CoronaStatsWorld world = newCoronaStats.getWorld();
+            if (world == null) {
+                updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR, "World stats not found");
+                return;
+            }
+
+            world.getChannelsStateMap().forEach(this::updateState);
+            countryListeners.forEach(listener -> listener.notifyOnUpdate(newCoronaStats));
+        }
+    }
+
+    public @Nullable CoronaStats getCoronaStats() {
+        return coronaStats;
     }
 }
