+++ bundles/org.openhab.binding.coronastats/src/main/resources/ESH-INF/thing/channel-type.xml
@@ -4,9 +4,51 @@
 	xmlns:thing="https://openhab.org/schemas/thing-description/v1.0.0"
 	xsi:schemaLocation="https://openhab.org/schemas/thing-description/v1.0.0 https://openhab.org/schemas/thing-description-1.0.0.xsd">
 
-	<channel-type id="count">
+	<channel-type id="cases">
 		<item-type>Number</item-type>
-		<label>Cases Count</label>
+		<label>Total Cases</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="today_cases">
+		<item-type>Number</item-type>
+		<label>New Cases</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="deaths">
+		<item-type>Number</item-type>
+		<label>Total Deaths</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="today_deaths">
+		<item-type>Number</item-type>
+		<label>New Deaths</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="recovered">
+		<item-type>Number</item-type>
+		<label>Recovered</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="active">
+		<item-type>Number</item-type>
+		<label>Active</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="critical">
+		<item-type>Number</item-type>
+		<label>Critical</label>
+		<state readOnly="true"/>
+	</channel-type>
+
+	<channel-type id="tests">
+		<item-type>Number</item-type>
+		<label>Tests</label>
 		<state readOnly="true"/>
 	</channel-type>
